feat: delete & edit caution confirmation messages

+++ app/src/main/java/com/devfiveurjc/agendaly/fragments/TaskInfoFragment.java
@@ -1,5 +1,6 @@
 package com.devfiveurjc.agendaly.fragments;
 
+import android.app.AlertDialog;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
@@ -48,17 +49,28 @@ public class TaskInfoFragment extends Fragment {
     public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
         this.binding.taskInfoDeleteButton.setOnClickListener(v -> {
-            // TODO: confirmation popup
-            CRUDTask.deleteTask(this.taskId);
-            NavHostFragment.findNavController(this)
-                    .navigate(R.id.action_TaskInfoFragment_to_TaskListFragment);
+            this.deleteCautionMessage();
         });
         this.binding.taskInfoEditButton.setOnClickListener(v -> {
             Bundle bundle = new Bundle();
-            bundle.putInt("taskId", taskId);
+            bundle.putInt("taskId", this.taskId);
             NavHostFragment.findNavController(this)
-                    .navigate(R.id.action_TaskInfoFragment_to_TaskEditFragment,bundle);
+                    .navigate(R.id.action_TaskInfoFragment_to_TaskEditFragment, bundle);
         });
+    }
+
+    private void deleteCautionMessage() {
+        AlertDialog.Builder alert = new AlertDialog.Builder(this.getContext());
+        alert.setMessage(getString(R.string.caution_text_delete) + getString(R.string.caution_advise))
+            .setCancelable(false)
+            .setTitle(R.string.caution)
+            .setPositiveButton(R.string.confirm, (dialogInterface, i) -> {
+                CRUDTask.deleteTask(this.taskId);
+                NavHostFragment.findNavController(this)
+                    .navigate(R.id.action_TaskInfoFragment_to_TaskListFragment);
+            })
+            .setNegativeButton(R.string.goback, (dialogInterface, i) -> dialogInterface.cancel());
+        alert.show();
     }
 
     @Override
